Move a file.
